Reject bandwidth limited values of N_{ref}
- Move code to for DSE on a single FPGA to a different C++ module
  (search_configuration.cpp)
- Refactor throughput calculation
- Reject bandwidth limited value at binary search.
